fix: rename use-checkbox-group
